requirements.txt
numpy~=1.17.2
opencv-contrib-python~=4.1.1.26
opencv-python~=4.1.1.26
Pillow~=6.2.0
